+++ src/Enterspeed.Delivery.Sdk/Api/Models/DeliveryApiResponse.cs
@@ -1,10 +1,12 @@
 ﻿using System.Net;
+using System.Net.Http.Headers;
 
 namespace Enterspeed.Delivery.Sdk.Api.Models
 {
     public class DeliveryApiResponse
     {
         public HttpStatusCode StatusCode { get; set; }
+        public HttpResponseHeaders Headers { get; set; }
         public string Message { get; set; }
         public DeliveryResponse Response { get; set; }
     }
@@ -12,6 +14,7 @@ namespace Enterspeed.Delivery.Sdk.Api.Models
     public class DeliveryApiResponse<T>
     {
         public HttpStatusCode StatusCode { get; set; }
+        public HttpResponseHeaders Headers { get; set; }
         public string Message { get; set; }
         public DeliveryResponse<T> Response { get; set; }
     }